+++ mods/BrewLAN_Gameplay/InvisibleWar/mod_info.lua
@@ -1,0 +1,13 @@
+name = "Invisible War"
+uid = "16678e1e-7fc9-11e5-8bcf-dayofsex0010"
+version = 1.0
+copyright = "© 2016–2018 Sean Wheeldon"
+description = "Everything has cloak. Did you remember your scouts? May behave weird with units that already have cloak."
+author = "Sean Wheeldon"
+icon = "/mods/BrewLAN_Gameplay/InvisibleWar/mod_icon.dds"
+selectable = true
+enabled = true
+exclusive = false
+ui_only = false
+before = {  }
+after = {  }