+++ mods/BrewLAN_Gameplay/Knives/mod_info.lua
@@ -1,0 +1,16 @@
+﻿name = "Knife Fight"
+uid = "25D57D85-9JA7-D842-BREW-KNIVES000002"
+version = 0.2
+copyright = "© 2007, 2017–2018 Sean Wheeldon" -- Mod created 2017, updated in 2018. Second knife asset made in 2007.
+description = "Units are armed with knives instead of guns. We have a knife fight on our hands lads."
+author = "Balthazar"
+icon = "/mods/BrewLAN_Gameplay/Knives/mod_icon.png"
+selectable = true
+enabled = true
+exclusive = false
+ui_only = false
+requires = { }
+requiresNames = { }
+conflicts = {}
+before = { }
+after = { }