+++ mods/BrewLAN_Gameplay/SuddenDeath/mod_info.lua
@@ -1,0 +1,16 @@
+name = "Sudden Death"
+uid = "25D57D85-9JA7-D842-BREW-SUDDEN000001"
+version = 0.1
+copyright = "© 2017–2018 Sean Wheeldon"
+description = "Units health reduced to between 2 and 20 health. Weapons damage values changed to 2."
+author = "Balthazar"
+icon = "/mods/BrewLAN_Gameplay/SuddenDeath/mod_icon.png"
+selectable = true
+enabled = true
+exclusive = false
+ui_only = false
+requires = { }
+requiresNames = { }
+conflicts = { }
+before = { }
+after = { }