+++ mods/BrewLAN_Gameplay/WaterGuard/mod_info.lua
@@ -1,0 +1,13 @@
+name = "Water Guard"
+uid = "16678e1e-7fc9-11e5-8bcf-waterguard10"
+version = 1.0
+copyright = "© 2016–2018 Sean Wheeldon"
+description = "Underwater things takes less damage from things that only hit the surface."
+author = "Sean Wheeldon"
+icon = "/mods/BrewLAN_Gameplay/WaterGuard/mod_icon.dds"
+selectable = true
+enabled = true
+exclusive = false
+ui_only = false
+before = {  }
+after = {  }